+++ service.d/start-lxc.sh
@@ -1,8 +1,0 @@
-lxc_start() {
-    # start lxc service
-    export HOME=/data/data/com.termux/files/home/.suroot
-    [ -x $PREFIX/bin/lxc-autostart ] &&
-        $PREFIX/bin/lxc-autostart -l DEBUG
-}
-
-lxc_start